// src/lib/interfaces/featured-playlists.ts
export interface ExternalUrls {
  spotify: string
}

export interface Image {
  height: number
  url: string
  width: number
}

export interface ExternalUrls2 {
  spotify: string
}

export interface Owner {
  display_name: string
  external_urls: ExternalUrls2
  href: string
  id: string
  type: string
  uri: string
}

export interface Tracks {
  href: string
  total: number
}

export interface Item {
  collaborative: boolean
  description: string
  external_urls: ExternalUrls
  href: string
  id: string
  images: Image[]
  name: string
  owner: Owner
  primary_color?: string
  public: boolean
  snapshot_id: string
  tracks: Tracks
  type: string
  uri: string
}

export interface Playlists {
  href?: string
  items: Item[]
  limit?: number
  next?: string
  offset?: number
  previous?: string
  total?: number
}

export interface IFeaturedPlaylists {
  message: string
  playlists: Playlists
}